supervisor: do not accidentally restart programs on update.
The problem is that when doing `supervisorctl update` all changed
programs are restarted.
We do not want to do that early and all at at the same time because it
may break our zero downtime strategy.

Therefore we update the programs right before starting them. This may
also trigger a start when autostart is enabled for the program, but this
is not a problem since we will just start the program anyway.