update supervisor config after update.
When buildout is ran, the supervisor.cfg may change. These changes need
to be loaded by the supervisor daemon.

There are two commands involved:

1. "bin/supervisorctl reread" tells us whether there are changes and
   what the changes are (e.g. program changed, added, removed).
2. "bin/supervisorctl update [programs..]" updates one, many or all
   programs, usually by restarting the program.

When we do a reread / update right after running buildout, it will
probably restart all instances at once, which violates the zero
downtime strategy. Therefore before ich "start" of a program, the
program is updated. At the end of the deployment most programs will
already be updated, but untracked programs are not and new programs are
not loaded.

In order to update all program changes we reread at the end of the
deployment and then update the changed programs one after another.
This may restart changed programs, it will start new programs and it
will shutdown removed programs.

However, we do never restart the zeo because this may result in
conection loss of the instances and this violates the zero downtime
strategy. This problem cannot be solved automatically and the maintainer
must restart the zeo at the right time (maintenance window).