Ensure precompilation when buildout is not run.
In environments where the service user (which runs the zope process)
is not allowed to write sources files it is necessary to precompile python
code and translation files on installation / update time.

When the precompile.cfg is used this is done in a complete buildout run.
If it isn't necessary to run buildout, we must still run the precompile
part in order to ensure that we have compiled changes in python code and
translation files which are shipped in the primary repository (source
checkout).